Move tests to shim.rs
